Created BaseManageRecipeFragment.kt and BaseManageRecipeViewModel.kt. Changed the Edit and Add Fragment and ViewModel correspondingly
