+++ app/src/main/java/com/fwrdgrp/recipesaving/ui/managerecipe/BaseManageRecipeFragment.kt
@@ -5,56 +5,175 @@ import androidx.fragment.app.Fragment
 import android.view.LayoutInflater
 import android.view.View
 import android.view.ViewGroup
+import android.widget.AdapterView
+import android.widget.ArrayAdapter
+import android.widget.TextView
+import androidx.core.content.ContextCompat
+import androidx.lifecycle.lifecycleScope
+import androidx.navigation.fragment.findNavController
+import androidx.recyclerview.widget.LinearLayoutManager
 import com.fwrdgrp.recipesaving.R
+import com.fwrdgrp.recipesaving.data.enums.Category
+import com.fwrdgrp.recipesaving.data.models.recipe.Ingredient
+import com.fwrdgrp.recipesaving.data.models.recipe.Instruction
+import com.fwrdgrp.recipesaving.data.models.recipe.Recipe
+import com.fwrdgrp.recipesaving.databinding.FragmentAddRecipeBinding
+import com.fwrdgrp.recipesaving.ui.adapters.AddIngredientAdapter
+import com.fwrdgrp.recipesaving.ui.adapters.AddInstructionAdapter
+import com.google.android.material.snackbar.Snackbar
+import kotlinx.coroutines.Dispatchers
+import kotlinx.coroutines.launch
 
-// TODO: Rename parameter arguments, choose names that match
-// the fragment initialization parameters, e.g. ARG_ITEM_NUMBER
-private const val ARG_PARAM1 = "param1"
-private const val ARG_PARAM2 = "param2"
+abstract class BaseManageRecipeFragment : Fragment() {
+    protected lateinit var binding: FragmentAddRecipeBinding
+    protected abstract val viewModel: BaseManageRecipeViewModel
 
-/**
- * A simple [Fragment] subclass.
- * Use the [BaseManageRecipeFragment.newInstance] factory method to
- * create an instance of this fragment.
- */
-class BaseManageRecipeFragment : Fragment() {
-    // TODO: Rename and change types of parameters
-    private var param1: String? = null
-    private var param2: String? = null
+    protected lateinit var instructionAdapter: AddInstructionAdapter
+    protected lateinit var ingredientAdapter: AddIngredientAdapter
+    protected lateinit var categoryAdapter: ArrayAdapter<Category>
 
-    override fun onCreate(savedInstanceState: Bundle?) {
-        super.onCreate(savedInstanceState)
-        arguments?.let {
-            param1 = it.getString(ARG_PARAM1)
-            param2 = it.getString(ARG_PARAM2)
-        }
-    }
+    protected val categories = Category.entries.toMutableList()
+    protected val selectedCategoryList = mutableListOf<Category>()
 
     override fun onCreateView(
         inflater: LayoutInflater, container: ViewGroup?,
         savedInstanceState: Bundle?
     ): View? {
-        // Inflate the layout for this fragment
-        return inflater.inflate(R.layout.fragment_base_manage_recipe, container, false)
+        binding = FragmentAddRecipeBinding.inflate(layoutInflater, container, false)
+        return binding.root
     }
 
-    companion object {
-        /**
-         * Use this factory method to create a new instance of
-         * this fragment using the provided parameters.
-         *
-         * @param param1 Parameter 1.
-         * @param param2 Parameter 2.
-         * @return A new instance of fragment BaseManageRecipeFragment.
-         */
-        // TODO: Rename and change types and number of parameters
-        @JvmStatic
-        fun newInstance(param1: String, param2: String) =
-            BaseManageRecipeFragment().apply {
-                arguments = Bundle().apply {
-                    putString(ARG_PARAM1, param1)
-                    putString(ARG_PARAM2, param2)
+    override fun onViewCreated(view: View, savedInstanceState: Bundle?) {
+        super.onViewCreated(view, savedInstanceState)
+        setupInstructionAdapter()
+        setupIngredientAdapter()
+        setupCategorySpinnerAdapter(categories, selectedCategoryList)
+        lifecycleScope.launch {
+            viewModel.finish.collect {
+                findNavController().popBackStack()
+            }
+        }
+        lifecycleScope.launch {
+            viewModel.error.collect {
+                showError(it)
+            }
+        }
+        binding.mbSubmit.setOnClickListener {
+            setupSubmitOnClick()
+        }
+    }
+
+    protected fun showError(msg: String) {
+        val snackbar = Snackbar.make(binding.root, msg, Snackbar.LENGTH_LONG)
+        snackbar.setBackgroundTint(ContextCompat.getColor(requireContext(), R.color.red)).show()
+    }
+
+    protected fun setupInstructionAdapter() {
+        instructionAdapter = AddInstructionAdapter(
+            mutableListOf(
+                Instruction(
+                    recipeId = 0,
+                    stepNumber = 1,
+                    description = ""
+                )
+            )
+        )
+
+        binding.rvAddInstruction.adapter = instructionAdapter
+        binding.rvAddInstruction.layoutManager = LinearLayoutManager(requireContext())
+        binding.ivAddInstruction.setOnClickListener {
+            instructionAdapter.addInstructions()
+        }
+    }
+
+    protected fun setupIngredientAdapter() {
+        ingredientAdapter = AddIngredientAdapter(
+            mutableListOf(
+                Pair(
+                    Ingredient(name = ""),
+                    Pair(0.0, "")
+                )
+            )
+        )
+
+        binding.rvAddIngredient.adapter = ingredientAdapter
+        binding.rvAddIngredient.layoutManager = LinearLayoutManager(requireContext())
+        binding.ivAddIngredient.setOnClickListener {
+            ingredientAdapter.addIngredient()
+        }
+    }
+
+    protected fun setupCategorySpinnerAdapter(
+        categoryList: MutableList<Category>,
+        selectedList: MutableList<Category>
+    ) {
+        binding.run {
+            categoryAdapter = ArrayAdapter(
+                requireContext(), android.R.layout.simple_spinner_item, categoryList
+            )
+            categoryAdapter.setDropDownViewResource(android.R.layout.simple_spinner_dropdown_item)
+            spCategory.adapter = categoryAdapter
+
+            spCategory.onItemSelectedListener = object : AdapterView.OnItemSelectedListener {
+                override fun onItemSelected(
+                    parent: AdapterView<*>?, view: View?, position: Int, id: Long
+                ) {
+                    val selected = categoryAdapter.getItem(position) ?: return
+                    if (selected == Category.INITIAL) return
+                    selectedList.add(selected)
+                    categoryList.remove(selected)
+                    val tvCategory = categoryChipDelete(categoryList, selectedList, selected)
+                    glCategory.addView(tvCategory)
+                    categoryAdapter.notifyDataSetChanged()
+                    spCategory.setSelection(0)
+                }
+
+                override fun onNothingSelected(parent: AdapterView<*>?) {}
+            }
+        }
+    }
+    protected fun categoryChipDelete(
+        categoryList: MutableList<Category>,
+        selectedList: MutableList<Category>,
+        item: Category,
+    ): TextView {
+        return binding.run {
+            TextView(requireContext()).apply {
+                text = item.name
+                setBackgroundResource(R.drawable.box_bg)
+
+                setOnClickListener {
+                    selectedList.remove(item)
+                    categoryList.add(item)
+                    categoryAdapter.notifyDataSetChanged()
+                    glCategory.removeView(this)
                 }
             }
+        }
     }
+
+    protected fun buildRecipe(category: List<Category>): Recipe {
+        binding.run {
+            return Recipe(
+                title = etTitle.text.toString(),
+                description = etDesc.text.toString(),
+                category = category,
+                estTime = etTime.text.toString().toInt(),
+                totalServing = etServing.text.toString().toInt(),
+                //Add when linkable
+                imageUri = ""
+            )
+        }
+    }
+
+    protected fun setupSubmitOnClick() {
+        lifecycleScope.launch(Dispatchers.IO) {
+            viewModel.submitRecipe(
+                buildRecipe(selectedCategoryList),
+                instructionAdapter.fetchInstructions(),
+                ingredientAdapter.fetchIngredient()
+            )
+        }
+    }
+
 }
